Update page layouts
I recently learned about manipulating page layout with tools like margin
and padding, and so I took advantage of that in this project to make
the pages easier and more pleasant to look at and read.

One major trick being used here is margin: auto. By setting a fixed
width to the whole page, and setting the left/right margin to auto, the
pages become not only centralized but also have a clean presentation.